Update patch set 1
Patch Set 1:

This:

  []int("hi") 

prints this error:

  cannot convert "hi" (type string) to type []int

on tip, int("hi") prints two errors:

  cannot convert "hi" to type int
  cannot convert "hi" (type string) to type int

In this patch you removed the second one, which is more informative and has the same format as the one we emit for []int("hi"). I was hoping to actually do the opposite (i.e. removing the first error, not the second one). 

The second one has also the same format of the error we emit for 

  s := "hi"
  _ = int(s)

which is:

  cannot convert s (type string) to type int

So yes, this patch fixes the 'double error' problem, but IMHO it would be better to fix it the other way (eliminating the less-informative less-consistent-with-the-rest first error).

Patch-set: 1
CC: Gerrit User 5385 <5385@62eb7196-b449-3ce5-99f1-c037f21e1705>