Update patch set 1
Patch Set 1:

I had originally intended to do this, but I'm unsure if it's possible.

The first error comes from convlit1 in const.go (cmd/compile/internal/gc/const.go#L411), if we try to replace the first `%v` with `%L` in that block, it results in:

    cannot convert "1" (type untyped string) to type int

To fix this, I tried assigning the type before printing the error (with defaultlitreuse, in that same block), and although it prints what we want now, it breaks a lot of other cases.

The second error comes from typecheck1 in typecheck.go (cmd/compile/internal/gc/typecheck.go#L1766), which calls convlit1. As far as I can tell, there's no way to omit the error in convlit1 and print it in this function instead.

Patch-set: 1
Reviewer: Gerrit User 21321 <21321@62eb7196-b449-3ce5-99f1-c037f21e1705>