Update patch set 1
Patch Set 1:

The if statement here is attempting to avoid the expense of ReadMemStats, because it did a full stop-the-world. Austin, ReadMemStats is cheap now, right? So we can probably just delete the if condition entirely and run the then block all the time, simplifying the code.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>