Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (3 comments)

How about just mention "too many", then?
It would be pointless to use constant, if we hard-cod value in the document (unless there's a mechanism to expanded it automatically), and actual number 10, 100, 1000 or 10000 doesn't really matter for the user, IMHO.
Even if it was just 2, if deemed too many, it's an error.

For the error message, we could possibly use something like this:
  panic(fmt.Sprintf("bufio.Scan: %s empty tokens without progressing", maxCo
nsecutiveEmptyReads))
I don't know if it worth introducing a new dependency though.

Patch-set: 2
Reviewer: Gerrit User 21690 <21690@62eb7196-b449-3ce5-99f1-c037f21e1705>