Update patch set 1
Patch Set 1:

(4 comments)

I think at long term the goal is to have an optimized block scheduling order and an optimized register allocator visiting order. So this sounds a good step.

If we experiment with block scheduling order and breaks the requirement, what do we do here? Fall back to reverse post order? Calling layoutOrder doesn't help then.


> Cherry may also be interested, in general.

Thanks. In fact this is related to a discussion David and I had in the office.

Patch-set: 1