encoding/base32: Exact DecodedLen() when using NoPadding
CL 47341 adds support for decoding non-padded messages. But DecodedLen() still
returns a multiple of 5 for messages without a padding, it is possible to
calculate the len exactly when using NoPadding.

This change makes DecodedLen() return the exact number of bytes that will be
written. A change to the decoding logic is also made so that it can handle
this case.

DecodedLen now has the same behaviour as DecodedLen in encoding/base64.

Fixes #20854

Change-Id: I729e0b1c0946c866fb675c854f835f366dd4b5a4