Update patch set 1
Patch Set 1:

I did this as a separate CL so we can decide which fix we want in Go 1.10. The "better" fix (this CL) turned out to be simpler than I expected, so maybe we just want to go ahead with this one.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>