Update patch set 1
Patch Set 1:

Thanks.

This looks OK for 1.9, but it doesn't fix #18968.  It should just update that issue.  #18968 will be fixed when we change the default for -buildmode=pie to use internal linking mode when possible, but we should not make that change for 1.9.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>