cmd/go: add go get -printdir
This flag will cause go get to print the path to where the package
already exists or where it will be installed. This is useful if GOPATH
contains multiple directories. If a package is in GOROOT the path to the
package in GOROOT will be returned.

The command accepts multiple packages as parameters, and the results
will be written to stdout on seperate lines.

Fixes #18119

Change-Id: I8cc5ab1e1b4b352380df682efa31b2d1b24d9752