Update patch set 5
Patch Set 5:

(2 comments)

See the 2 inline comments.

> I'm also slightly confused as to why Dmitri submitted the issue instead of just a CL.

See https://github.com/golang/go/issues/21026#issuecomment-315557620.

> Why not fix all three at once, as well?

I believe the scope of the CL is to tackle only one of the 3, because there were some challenges with doing the other 2 and I suggested to Iccha to just focus on one for now. That's why it says "Helps #21026." rather than "Fixes #21026." Iccha can confirm whether she'd like to proceed as is, or expand the scope of CL to fix the other 2 places as well.

Patch-set: 5
CC: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>