net/http: add client.go and fix failing benchmark
The current implementation of httptest.Server considers a function
Client that returns an http.Client for client tests on a server. While
this is fine for most tests, it can cause failures due to not having
enough available file descriptors. The client transport is not cleaned
from idle transport connections until the test function finishes and the
server gets closed (only after running all parallel tests).

Following the httptest.Server design, this commit adds an
httpserver.Client struct that has a CloseIdleTransport method that a
caller can use in order to close all idle connections in the client
transport.

Fixes #20166

Change-Id: Idff2c3c76f5ba0ec26c32a4fd26fb50383d298a3