Update patch set 2
Patch Set 2:

Is this Go-specific or is this a Linux/namespace/container lesson in the syscall package?

Normally we don't handhold too much down in these guts unless it's a Go-specific surprise or a popular trap.

I don't know enough to know what the case is here.

I'm fine with this if you are.

Patch-set: 2
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>