runtime: always use 1MB stacks on 32-bit Windows
Commit c2c07c7989 changed the linker and runtime to always use 2MB
stacks on 64-bit Windows. This is the corresponding change to make
32-bit Windows always use large (1MB) stacks because it's difficult to
detect when Windows applications will call into arbitrary C code that
may expect a large stack.

This is done as a separate change because it's possible this will
cause too much address space pressure for a 32-bit address space. On
the other hand, cgo binaries on Windows already use 1MB stacks and
there haven't been complaints.

Fixes #20975.

Change-Id: I8ce583f07cb52254fb4bd47250f1ef2b789bc490