net/http: improve signature of Redirect, NewRequest
In CL https://golang.org/cl/4893043, a new package named "url" was
created (it is currently known as "net/url"). During that change,
some identifier name collisions were introduced, and two parameters
in net/http were renamed to "urlStr". This name is suboptimal for
godoc purposes.

Since that change 6 years ago, Go has continued to put high emphasis
on the quality and readability of the documentation. Sometimes, that
means making small sacrifices in the implementation details of a package
to ensure that the godoc reads better, since that's what the majority
of users interact with. See https://golang.org/s/style#named-result-parameters:

> Clarity of docs is always more important than saving a line or two in your function.

During the review of https://golang.org/cl/4893043, it was noted
by @rsc that having to rename url parameters to urlStr was suboptimal:

> It's unfortunate that naming the package url means
> you can't have a parameter or variable named url.

However, at the time, the name of the url package was still being
decided, and uri was an alternative name under consideration.
The reason urlStr was chosen is because it was a lesser evil
compared to naming the url package uri instead:

> Let's not get hung up on URI vs. URL, but I'd like s/uri/urlStr/ even for just
> that the "i" in "uri" looks very similar to the "l" in "url" in many fonts.

> Please let's go with urlStr instead of uri.

Now that we have the Go 1 compatibility guarantee, the name of the
net/url package is fixed. However, it's possible to improve the
signature of Redirect, NewRequest functions in net/http package
for godoc purposes by moving the implementation into unexported
helpers, and renaming urlStr parameter to url in the exported funcs.
This CL does so.

Updates #21077.

Change-Id: Ibcc10e3825863a663e6ad91b6eb47b1862a299a6