net/http: improve signature of Redirect, NewRequest
The goal of this this change is to try to improve the readability
of godoc documentation of net/http for the two mentioned funcs.

DO NOT SUBMIT, this is just a prototype for review and discussion.

I wanted to try a prototype of the idea for an alternative fix I
suggested in https://github.com/golang/go/issues/21077#issuecomment-316205216.

There are two ways I found of being able to expose a nicer `url string`
parameter in godoc of net/http package. This is the first way, by
moving the implementation into an unexported helper. The downside
is an extra call.

Another way I found is to move the two methods into a separate file
and do:

	import urlpkg "net/url"

In that file only (to minimize disruption elsewhere). That seemed
more invasive of a change, but viable, and does not incur an extra
call. However, it requires moving Redirect into a separate file,
which can seem pretty arbitrary and unfortunate.

The question I want to answer is... Is doing this worth it?
It's in spirit of Go to go out of the way to make sure the
documentation is as nice as it can be, but is this minor improvement
worth the cost of extra complexity/unpleasantness in the implementation?

Updates #21077.

Change-Id: Ibcc10e3825863a663e6ad91b6eb47b1862a299a6