cmd/link: don't use dynreloc in external linking mode
Currently, dynreloc isn't called in external linking mode when the object
format is ELF. Although dynreloc is called in external linking mode when
the object format is PE or Mach-O, I suppose it doesn't do anything
important in that case. This CL add early return to make sure the behavior.
So we could simplify rest of code.

Change-Id: I41dc5de5a16fb4dd0685a71ece284995ab62d8ae