Update patch set 2
Patch Set 2: Code-Review-1

I'm -1 on switching to h4. On GitHub it is rendered with a font which has the same size of the normal text, which means that the titles appear just to be random strings in bold. 

This makes it actually harder to read, especially if you're scanning through a long submission (containing a long pasted text) to find a section you care about in the report.

Patch-set: 2
Reviewer: Gerrit User 5340 <5340@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1