Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Since this is an optimization change, you should add benchmark results to the commit message.
> 
> Run every math/big benchmark touched by this change multiple times (e.g. with
> go test .. -count 8), save old results and new results in two different files, and then paste the output of $ benchstat old.txt new.txt in the commmit message.


I wasn't sure the process for reporting benchmarks and thought that would be a very long commit message with all the results, so I uploaded before and after here:

https://perf.golang.org/search?q=upload:20170726.2

There wasn't a ModInverse benchmark, so I created one that copies the GCD benchmark.  Let me know if I should update the CL including that additional benchmark and pasting those full results into the commit message.  

From the linked benchstat, the gain for GCD is ~20% when the arguments have similar length but can degrade up to ~20% for very unbalanced arguments e.g. GCD10x10000/WithXY (1 Word, 157 Words) where the trade-off is multiplication, addition work in the loop versus the final multiplication, addition and division.  The gain for ModInverse diminishes for unbalanced arguments but is always a win in the range measured, ~25% for balanced inputs degrading to not statistically significant at the most unbalanced.

This probably needs a decision on whether the change is useful generally since there is that degradation in the unbalanced GCD benchmarks.  I would think it would depend on the typical range of inputs encountered.

Patch-set: 1
Reviewer: Gerrit User 20650 <20650@62eb7196-b449-3ce5-99f1-c037f21e1705>