math/big: fix roundShortest overflow
Float.prec is an uint32, and when prec is big.MaxPrec (aka MaxUint32),
an (x.prec + 1) computation in roundShortest overflows, causing havoc
in a later nat.sub call.

We could fix this by switching to int64s during the computation, but
it doesn't really matter because the actual maximum precision is
limited by the amout of available memory, and for a prec near 2**32
the algorithm roundShortest uses will allocate so much memory in the
following mantissa shifts that the computation will never end.

For this reason, simply falling back to x.prec when x.prec+1 overflows
fixes the issue without changing the observable behaviour of the
roundShortest function.

Fixes #20867

Change-Id: I202f2fb25d1cb02bf81c4516cd05461836e96526