math/big: fix roundShortest overflow
Float.prec is an uint32, and when prec is big.MaxPrec (aka MaxUint32),
an (x.prec + 1) computation in roundShortest overflows, causing havoc
in a later nat.sub call.

Simply falling back to x.prec when x.prec+1 overflows is a best-effort
strategy that fixes the issue and is unlikely to change the observable
behaviour of the roundShortest function.

Fixes #20867

Change-Id: I202f2fb25d1cb02bf81c4516cd05461836e96526