time: use RFC1123Z in test to force numeric zone abbreviation
The testZoneAbbr assumes that

  Parse(RFC1123, t1.Format(RFC1123))

will always work, but this is not true because Format will fall back
to the numeric timezone (like -07) when the current timezone has no
invented abbreviation (like BRT), but Parse won't accept the numeric
timezone when the layout specifies 'MST' (the abbreviation).

To fix this, change the test to use RFC1123Z, which is like RFC1123
except the timezone is always represented by the numeric zone, which
is guaranteed to work (unlike timezone abbreviations, which are being
phased out by IANA for certain timezone).

Fixes #21183

Change-Id: If04691cc23ae1075d8a953733024e17f5a7646de