encoding/csv: report line start line in errors
Errors returned by Reader currently the line where the Reader originally
encountered the error. The problem is that this line does not always
correspond with the line the current record/field started at.

This can easily happen with LazyQuotes as seen in #19019, but also
happens for example when a quoted fields has no closing quote and
the parser hits EOF before it finds another quote.

When this happens finding the erroneous field can be somewhat
complicated and time consuming, so n most cases it would be better to
report the line where the record started.

This change updates Reader to keep track of the line on which a record
begins and use this for errors instead of the current line, making it
easier to find errors.

Although a user-visible change, this should have no bug impact on existing
code, since most users don't explicitly work with the line in the error
and probably already expect this behaviour.

Updates #19019

Change-Id: Ic9bc70fad2651c69435d614d537e7a9266819b05