NOT FOR SUBMISSION runtime: change selectdone to epoch
selectdone was a very subtle construct (as evidenced by the 55 line
comment). Simplify by using an incrementing counter stored in the
G of the enqueued process and store the counter value in the sudog
when enqueuing.

This gets rid of a cross goroutine stack write and a systemstack
call. It also lays the groundwork for treating sudogs as any other
heap structure that doesn't point into stack.

Change-Id: Ia32acb33fd3e0d554c699b97d670fccbefd9eb68