Update patch set 1
Patch Set 1:

NOTE: I'm not entirely sure that the epoch is necessary and it might be that a flag within the G will do. Once a goroutine is running, it clears itself out of all the queues that it's in, and if it clears the flag afterwards, there shouldn't be any potential for an ABA problem. I might be missing something here and the construct with selectdone seems almost too convoluted for this to not matter.

Patch-set: 1
Reviewer: Gerrit User 5310 <5310@62eb7196-b449-3ce5-99f1-c037f21e1705>