Update patch set 5
Patch Set 5:

I was thinking how to test this.

Just testing that poll.Init works does not provide great confidence in me. I was thinking about creating real net sockets and files and checking these. But I need access to internal bits in 2 different packages net and internall/poll, or os and internall/poll, and that is not possible. Suggestions are welcome.

Thank you

Alex

Patch-set: 5