cmd/compile: don't use map{assign,delete}_fast* for pointer-ful types
The fast map modification functions don't perform write barriers when
writing or zeroing the key in the map, so they aren't safe when the
key contains pointers.

This commit implements a quick fix for this: simply disable the fast
map functions for pointer-ful key types. An alternate solution would
be for the fast map functions to inspect the key type an perform a
write barrier, but disabling them is clearly safer for a last-minute
fix before the release.

The fast map functions were introduced during Go 1.9 development, so
this is a new bug.

Fixes #21297.

Change-Id: I756704dfd84025e407879b66efb4e773a811fa68