Update patch set 5
Patch Set 5:

While, this is irrefutably a step in the right direction - I just wanted to add a concern on one of the points on the commit, to ensure that this doesn't mark an end to the timer resolution pursuit on Windows. 

"When the process is non-idle, it's already consuming CPU so it doesn't really matter if the OS also takes timer interrupts more frequently."

I believe this statement is highly misguided. While go programs generally tend to be server programs which have the luxury to be seen as the primary process on the system/container - it generally isn't the case on Windows. They are more often than not, general purpose programs. And it's common in the Windows world to have 2k-3k threads sitting idle, depending on the system. I tend to have a clean system with rather obsessive control over the background services, and programs, and I have 1959 threads right now. While the Windows Scheduler is quite efficient at it, that's still a lot of book-keeping. So such a statement is quite harmful here. Because, now since the Windows timer resolution has been brought down, Windows now has to do all it's book-keeping every 1ms, instead of every 15ms. And the Go program is not the only program running in most desktop uses cases. So, being a good citizen is important, and can't just be ignorant of it. A Go programs that's sleeps and uses CPU in small frequent bursts will still do a lot of harm.

This statement is only true when you consider server scenarios. But I think it would be a safe guess to say that's not as much relevant on Windows, as much as desktop scenarios. 

PS: The case that Chrome does similar things also doesn't make it right - That's one of the reasons why Edge browser is far more efficient and uses 30-40% less battery than Chrome as per Microsoft's claim. Chrome simply seems to raise the timer, even when you just move the mouse around. Not to mention Chrome is an application, while Go is a language - so it's more important to get things right.

Patch-set: 5
CC: Gerrit User 23868 <23868@62eb7196-b449-3ce5-99f1-c037f21e1705>