Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > While, this is irrefutably a step in the right direction - I just wanted to add a concern on one of the points on the commit, to ensure that this doesn't mark an end to the timer resolution pursuit on Windows. 
> > 
> > "When the process is non-idle, it's already consuming CPU so it doesn't really matter if the OS also takes timer interrupts more frequently."
> > 
> > I believe this statement is highly misguided. While go programs generally tend to be server programs which have the luxury to be seen as the primary process on the system/container - it generally isn't the case on Windows. They are more often than not, general purpose programs. And it's common in the Windows world to have 2k-3k threads sitting idle, depending on the system. I tend to have a clean system with rather obsessive control over the background services, and programs, and I have 1959 threads right now. While the Windows Scheduler is quite efficient at it, that's still a lot of book-keeping. So such a statement is quite harmful here. Because, now since the Windows timer resolution has been brought down, Windows now has to do all it's book-keeping every 1ms, instead of every 15ms. And the Go program is not the only program running in most desktop uses cases. So, being a good citizen is important, and can't just be ignorant of it. A Go programs that's sleeps and uses CPU in small frequent bursts will still do a lot of harm.
> > 
> > This statement is only true when you consider server scenarios. But I think it would be a safe guess to say that's not as much relevant on Windows, as much as desktop scenarios. 
> 
> I'm afraid I don't understand your argument. I don't think I was assuming anything about running in a server environment or not having co-tenants.
> 
> The argument I was making is pretty simple: from a power consumption stand-point, if the CPU is idling in a low-power state, you want to leave it idling. But if the CPU is already running your process, then taking more frequent interrupts won't make power consumption worse.
> 
> Can you say what specific "harm" is caused by a Go program that uses CPU in small frequent bursts?
> 
> (Also, I would hope that a large number of idle threads would have almost no effect on the scheduler's CPU consumption. I'm rather disappointed if that's not the case.)

It is correct that idle threads have no effect on the scheduler's consumption. But only "idle" threads don't. My argument is that, the assumption that if the CPU is already running your process, you can let the OS take frequenct interrupts isn't correct in all environments. 

In desktop environments, it is quite common for a threads of other programs to be active. For the sake of the argument, say a program like "Process Explorer", or a browser like Edge that's already taking constant interrupts, but not switching resolution timers to 1ms. And now, Go's runtime puts everything on overdrive. Small bursts of CPU by the Go program, will let all have the kernel doing a lot more work.  

While the grand solution is possibly the UMS scheduler, and I do agree that this is a great improvement over how it was. As I said in the very first line "I just wanted to add a concern on one of the points on the commit, to ensure that this doesn't mark an end to the timer resolution pursuit on Windows." :)

Patch-set: 5